+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 #### exonum
 
+- `NotFound` error message for `explorer/v1/block` endpoint now includes
+  the actual blockchain height. (#1498)
+
 - `system/v1/rebroadcast` endpoint has been removed. (#1445)
 
 - Added a possibility to specify compression algorithm for the database. (#1447)